# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /readyz:
    get:
      operationId: readinessCheck
      tags: [health]
      summary: Database and migration readiness
      description: Returns 503 while draining or if the database/security/catalogue schema is unavailable. Does not verify external providers or complete business workflows.
      responses:
        "200":
          description: Ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
        "503":
          description: Not ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /admin/accounts:
    get:
      operationId: listManagedAccounts
      tags: [accounts]
      summary: List buyer and seller accounts (manager only)
      security: [{ sessionCookie: [] }]
      parameters:
        - {
            in: query,
            name: page,
            schema: { type: integer, minimum: 0, maximum: 100000, default: 0 },
          }
        - { in: query, name: q, schema: { type: string, maxLength: 120 } }
        - {
            in: query,
            name: status,
            schema:
              { type: string, enum: [all, active, suspended], default: all },
          }
      responses:
        "200":
          description: Up to 25 accounts, ordered by join date and ID descending
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ManagedAccountPage" }
        default:
          description: Invalid request, missing/unauthorized session or unavailable service
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /admin/accounts/summary:
    get:
      operationId: getManagedAccountSummary
      tags: [accounts]
      summary: Count buyer and seller accounts (manager only)
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Current account counts
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ManagedAccountSummary" }
        default:
          description: Missing/unauthorized session or unavailable service
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /admin/accounts/{id}/status:
    post:
      operationId: changeManagedAccountStatus
      tags: [accounts]
      summary: Suspend or reactivate a buyer/seller account (manager only)
      description: Staff/self changes are forbidden. Status change, session revocation and audit insertion are atomic. A stale version or redundant change returns 409; refresh rather than repeating blindly. Only a successful transition writes an audit record.
      security: [{ sessionCookie: [] }]
      parameters:
        - {
            in: path,
            name: id,
            required: true,
            schema: { type: string, format: uuid },
          }
        - {
            in: header,
            name: X-Fotizo-Request,
            required: true,
            schema: { type: string, enum: ["1"] },
          }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AccountStatusChange" }
      responses:
        "200":
          description: Account state changed and all previous sessions revoked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountStatusChangeResult" }
        default:
          description: 400 invalid input; 401 expired session; 403 forbidden; 404 missing account; 409 stale version; 503 temporary failure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /admin/account-audit:
    get:
      operationId: listAccountAudit
      tags: [accounts]
      summary: Read account status history (manager only)
      security: [{ sessionCookie: [] }]
      parameters:
        - {
            in: query,
            name: page,
            schema: { type: integer, minimum: 0, maximum: 100000, default: 0 },
          }
        - { in: query, name: targetId, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Up to 25 audit records, most recent first; names are current account names
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountAuditPage" }
        default:
          description: Invalid filters, missing/unauthorized session or unavailable service
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /products:
    get:
      operationId: listCatalogueProducts
      tags: [catalogue]
      summary: Browse a bounded page of published products
      description: Filters and sorting apply to the whole channel. Pages are zero-based, ordered with deterministic creation-time/ID tie breakers. Offset pages can shift during concurrent catalogue changes. Prices are GBP display prices. Unknown query fields and reversed price ranges return 400.
      parameters:
        - {
            in: query,
            name: channel,
            schema:
              { type: string, enum: [marketplace, shop], default: marketplace },
          }
        - {
            in: query,
            name: page,
            schema: { type: integer, minimum: 0, maximum: 10000, default: 0 },
          }
        - {
            in: query,
            name: pageSize,
            schema: { type: integer, minimum: 1, maximum: 48, default: 24 },
          }
        - { in: query, name: q, schema: { type: string, maxLength: 120 } }
        - {
            in: query,
            name: category,
            schema: { type: string, minLength: 1, maxLength: 80 },
          }
        - {
            in: query,
            name: sort,
            schema:
              {
                type: string,
                enum: [newest, price-asc, price-desc, rating, discount, best-selling],
                default: newest,
              },
          }
        - {
            in: query,
            name: minPrice,
            schema: { type: number, minimum: 0, maximum: 99999999.99 },
          }
        - {
            in: query,
            name: maxPrice,
            schema: { type: number, minimum: 0, maximum: 99999999.99 },
          }
        - {
            in: query,
            name: minRating,
            schema: { type: number, minimum: 0, maximum: 5 },
          }
        - { in: query, name: inStock, schema: { type: boolean } }
        - { in: query, name: discounted, schema: { type: boolean } }
      responses:
        "200":
          description: Page and matching total from the same database snapshot
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CataloguePage" }
        default:
          description: Invalid filters or service failure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /products/categories:
    get:
      operationId: listCatalogueCategories
      tags: [catalogue]
      summary: Published product counts by category for the entire channel
      parameters:
        - {
            in: query,
            name: channel,
            schema:
              { type: string, enum: [marketplace, shop], default: marketplace },
          }
      responses:
        "200":
          description: Category keys, counts and representative images; includes out-of-stock listings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CatalogueCategory" }
        default:
          description: Invalid filters or service failure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: fotizo_token
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string }
    ManagedAccount:
      type: object
      required: [id, name, email, role, createdAt, suspendedAt, statusVersion]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [buyer, seller] }
        createdAt: { type: string, format: date-time }
        suspendedAt: { type: [string, "null"], format: date-time }
        statusVersion: { type: integer, minimum: 0 }
    ManagedAccountPage:
      type: object
      required: [items, page, hasMore]
      properties:
        items:
          {
            type: array,
            items: { $ref: "#/components/schemas/ManagedAccount" },
          }
        page: { type: integer }
        hasMore: { type: boolean }
    ManagedAccountSummary:
      type: object
      required: [total, active, suspended]
      properties:
        total: { type: integer, minimum: 0 }
        active: { type: integer, minimum: 0 }
        suspended: { type: integer, minimum: 0 }
    AccountStatusChange:
      type: object
      additionalProperties: false
      required: [action, reason, expectedVersion]
      properties:
        action: { type: string, enum: [suspend, reactivate] }
        reason: { type: string, minLength: 10, maxLength: 1000 }
        expectedVersion: { type: integer, minimum: 0, maximum: 2147483646 }
    AccountStatusChangeResult:
      type: object
      required: [id, suspendedAt, statusVersion, auditId]
      properties:
        id: { type: string, format: uuid }
        suspendedAt: { type: [string, "null"], format: date-time }
        statusVersion: { type: integer }
        auditId: { type: string, format: uuid }
    AccountAuditEntry:
      type: object
      required:
        [
          id,
          actorId,
          actorName,
          targetUserId,
          targetName,
          action,
          reason,
          statusVersion,
          createdAt,
          previousSuspendedAt,
          suspendedAt,
        ]
      properties:
        id: { type: string, format: uuid }
        actorId: { type: string, format: uuid }
        actorName: { type: string }
        targetUserId: { type: string, format: uuid }
        targetName: { type: string }
        action: { type: string, enum: [suspend, reactivate] }
        reason: { type: string }
        statusVersion: { type: integer }
        createdAt: { type: string, format: date-time }
        previousSuspendedAt: { type: [string, "null"], format: date-time }
        suspendedAt: { type: [string, "null"], format: date-time }
    AccountAuditPage:
      type: object
      required: [items, page, hasMore]
      properties:
        items:
          {
            type: array,
            items: { $ref: "#/components/schemas/AccountAuditEntry" },
          }
        page: { type: integer }
        hasMore: { type: boolean }
    CatalogueProduct:
      type: object
      required:
        [
          id,
          channel,
          status,
          title,
          description,
          price,
          originalPrice,
          rating,
          reviewCount,
          seller,
          sellerId,
          category,
          image,
          images,
          inStock,
          stockCount,
          tags,
          specs,
        ]
      properties:
        id: { type: string, format: uuid }
        channel: { type: string, enum: [marketplace, shop] }
        status: { type: string, enum: [active, unpublished] }
        title: { type: string }
        description: { type: string }
        price: { type: number }
        originalPrice: { type: [number, "null"] }
        rating: { type: number }
        reviewCount: { type: integer }
        seller: { type: string }
        sellerId: { type: string, format: uuid }
        category: { type: string }
        image: { type: string }
        images: { type: array, items: { type: string } }
        inStock: { type: boolean }
        stockCount: { type: integer }
        tags: { type: array, items: { type: string } }
        specs: { type: object, additionalProperties: { type: string } }
    CataloguePage:
      type: object
      required: [items, total, page, pageSize, hasMore]
      properties:
        items:
          {
            type: array,
            items: { $ref: "#/components/schemas/CatalogueProduct" },
          }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 0 }
        pageSize: { type: integer, minimum: 1, maximum: 48 }
        hasMore: { type: boolean }
    CatalogueCategory:
      type: object
      required: [category, count, image]
      properties:
        category: { type: string }
        count: { type: integer, minimum: 1 }
        image: { type: string }