# lib/api-zod/src/generated/types/accountAuditPage.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AccountAuditEntry } from "./accountAuditEntry";

export interface AccountAuditPage {
  items: AccountAuditEntry[];
  page: number;
  hasMore: boolean;
}