# lib/api-zod/src/generated/types/managedAccountPage.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ManagedAccount } from "./managedAccount";

export interface ManagedAccountPage {
  items: ManagedAccount[];
  page: number;
  hasMore: boolean;
}