# lib/api-zod/src/generated/types/managedAccountSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ManagedAccountSummary {
  /** @minimum 0 */
  total: number;
  /** @minimum 0 */
  active: number;
  /** @minimum 0 */
  suspended: number;
}